add angular for frontend
the backend was modified to expose fully rest controllers which return
json, so that angular can be used as a fully separated component for the
frontend

+++ src/main/java/it/tino/blog/article/ArticleSpringDataSource.java
@@ -42,8 +42,12 @@ public class ArticleSpringDataSource implements ArticleRepository {
     }
 
     @Override
-    public void deleteById(UUID id) {
-        articleDao.deleteById(id);
+    public boolean deleteById(UUID id) {
+        if (articleDao.existsById(id)) {
+            articleDao.deleteById(id);
+            return true;
+        }
+        return false;
     }
 
     private SpringArticle domainToDb(Article article) {
